fix bugs

+++ src/services/carts/deleteBookOnCart.service.ts
@@ -3,6 +3,8 @@ import { Books_Cart } from "../../entities/books_cart.entity"
 import { AppError } from "../../errors"
 
 const deleteBookOnCartService = async (bookCartId:string, userId:string): Promise<object> =>{
+    const bookCartRepository = AppDataSource.getRepository(Books_Cart)
+
 
     if(!bookCartId){
         throw new AppError('Bad data', 400)
@@ -10,19 +12,6 @@ const deleteBookOnCartService = async (bookCartId:string, userId:string): Promis
 
     if(!userId){
         throw new AppError('Bad data', 400)
-    }
-
-    const bookCartRepository = AppDataSource.getRepository(Books_Cart)
-
-    const findCart = await bookCartRepository.createQueryBuilder('books_cart')
-    .innerJoin("books_cart.cart", "cart")
-    .where("cart.user = :id_user", {id_user: userId})
-    .andWhere("books_cart.id = :id", { id : bookCartId})
-    .select("*")
-    .getOne()
-
-    if(!findCart){
-        throw new AppError('Unauthorized', 403)
     }
 
     const bookCart = await bookCartRepository.findOneBy({
@@ -33,12 +22,23 @@ const deleteBookOnCartService = async (bookCartId:string, userId:string): Promis
         throw new AppError('Not found', 404)
     }
 
+    /* const findCart = await bookCartRepository.createQueryBuilder('books_cart')
+    .innerJoin("books_cart.cart", "cart")
+    .where("cart.user = :id_user", { id_user: userId })
+    .andWhere("books_cart.id = :id", { id : bookCartId })
+    .select("*")
+    .getOne()
+
+    if(!findCart){
+        throw new AppError('Unauthorized', 403)
+    } */
+
     await bookCartRepository.delete({
         id:bookCartId
     })
 
 
-    return ({message : "Cart iten deleted"})
+    return
 }
 
 export default deleteBookOnCartService
